+++ src/main/java/com/dailyon/productservice/reviewaggregate/entity/ReviewAggregate.java
@@ -23,12 +23,12 @@ public class ReviewAggregate {
     private Product product;
 
     @Column(nullable = false)
-    private Float avgRating;
+    private Double avgRating;
 
     @Column(nullable = false)
     private Long reviewCount;
 
-    public static ReviewAggregate create(Product product, Float avgRating, Long reviewCount) {
+    public static ReviewAggregate create(Product product, Double avgRating, Long reviewCount) {
         return ReviewAggregate.builder()
                 .product(product)
                 .avgRating(avgRating)
